gallium: remove PIPE_CAP_BUFFER_SHARE cap and get_param sw_winsys hook
The kms_swrast driver has a separate InitScreen hook for its DriverAPI
from the rest of the DRI2 drivers, all of which capable of buffer
sharing. As such we no longer need to dive through the pipe-driver and
winsys layers in order to determine if the driver can share buffers or
not and we can explicitly set screen->can_share_buffer in InitScreen.

XXX: Squash with the original commit ?

Cc: Giovanni Campagna <gcampagna@src.gnome.org>
Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>